Start of a Kafka Streams powered websocket system
Why:
For fun. There is a real challenge to getting events from Kafka to a
websocket connection at scale

How:
Combination of kafka streams to produce the course grain
partitioning. This partitioning is used by the websocket clients to
connect to the server for their target course grain partition.

Then use event-ruler to provide fine grain subscriptions. This allows
the websocket to specify the small subset of events the connection is
actually interested in