Test to demonstrate State restoration works
Why:
Because of it is using Mutable Collections -- HashMap and ArrayList --
for the state *and* jackson isn't written to restore it to those
collection types, need to demonstrate that collections are recreated
copies of the data serialized to disk

How:
Starts a cluster and a partitioner. Then it sends a input record and
awaits for it make it through the cluster. Then the partitioner is
closed and a new one is created.

To trigger the notification loop, a new record is sent and this
results in the first record being published to the new subscription