Build out the streaming application for the queues
Why:
The previous application was over complex and broken -- it would send
events repeatedly to the queue.

How:
This code uses the `selectKey` method to rekey the input topic based
on the json path passed into the partitioner. Then it processes each
event and finds any matches and publishes to their respective queues