Pipe the kafka output to the web service
Why:
This attempts to connect the running streams application to the web
service. When a user submits a JSON filter to the enpdoint, create a
subscription by passing the rules to the subscription service

How:
Creates a new interface `SubscriptionService` that holds the
subscribe and unsubscribe methods. These are called by the events
handler based on the current state of the http3 connection