Messy commit to hook events into the system
Why:
This commit demostrates how to mock the `SubscriptionService` in the
integration test. This allows the event to be pushed from the kafka
streams to the http3 connection.

How:
Creates a thread to run and check the count down latch for
completion. The thread then polls the queue from the subscription and
flushes the event to the http3 connection queue.