Query Kafka Metadata to locate host for a keybytes
Why:
This method will be used by the frontend to ask the partitioner "Which
host has data for this key".

How:
First query the topic to get the number of partitions and use that to
query the default partitioner for the partition. Then query the
consumer group to find the host assigned to that partition