Messy WIP on connecting the web front end to the location
Why:
The web frontend should query the subscription service for the host
location of the data. This will instruct the Http3 client to connect
to a different host all together to get events.

What is wrong:
Given a URI returned we need to construct the proper Http3
response. not sure that is what is happening so far in this code.